// frontend/src/app/profile-content/update-interface.ts
export interface UpdateData {
    id: string;
    profileImage: string;
    firstName: string;
    middleName: string;
    lastName: string;
    email: string;
    mobileNumber: string;
    whatsAppNumber: string;
    lineOne: string;
    lineTwo: string;
    streetName: string;
    state: string;
    country: string;
    zipCode: string;
    collegeName: string;
    courseName: string;
    graduationYear: string;
    certificateName: string;
    certificateCourseName: string;
    certifiedYear: string;
    companyName: string;
    jobRole: string;
    duration: string;
    dateOfJoining: string;
    dateOfReleaving: string;
    noticePeriod: string;
    declaration: string;
    creator: string;
}